requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-phonenumber-field==7.0.2
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-extra-fields==3.4.1
idna==3.4
phonenumbers==8.13.4
Pillow==9.4.0
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14

aiogram~=2.25.1